fix: updating references in index.html to be absolute
index.html is the content delivered when a page is not found. If someone visits a page that doesn't exists for example mariapretiz.com/compras/ then the content of index.html is delivered and the path in the browser is preserved. Since the path is preserved, having the javascript files being relative instead of absolute makes the browser try to load a file from /compras/js/jquery.js instead of /js/jquery.js